+++ src/mcmc/hmc.jl
@@ -454,17 +454,9 @@ function AHMCAdaptor(
     pc = AHMC.MassMatrixAdaptor(metric)
     da = AHMC.StepSizeAdaptor(alg.δ, ϵ)
 
-    if iszero(alg.n_adapts)
-        adaptor = AHMC.Adaptation.NoAdaptation()
-    else
-        if metric == AHMC.UnitEuclideanMetric
-            adaptor = AHMC.NaiveHMCAdaptor(pc, da)  # there is actually no adaptation for mass matrix
-        else
-            adaptor = AHMC.StanHMCAdaptor(pc, da)
-            AHMC.initialize!(adaptor, nadapts)
-        end
-    end
-
+    iszero(alg.n_adapts) && return AHMC.Adaptation.NoAdaptation()
+    adaptor = AHMC.StanHMCAdaptor(pc, da)
+    AHMC.initialize!(adaptor, nadapts)
     return adaptor
 end